Issue 3 (additional thread for results output); Issue 6 (implement calculation in multi-threaded mode);

+++ trunk/sudoku/sudoku/SudokuSolution.cs
@@ -1,20 +1,18 @@
-using System.Collections.Generic;
-
 namespace sudoku
 {
     public class SudokuSolution
     {
         static object lockObject = new object();
-        public const int NO_SOLUTION = 0;
-        public const int UNIQUE_SOLUTION = 1;
-        public const int MULTIPLE_SOLUTION = 2;
+        public const byte NO_SOLUTION = 0;
+        public const byte UNIQUE_SOLUTION = 1;
+        public const byte MULTIPLE_SOLUTION = 2;
 
         readonly int taskId;
         readonly SudokuTable initialTable;
-        int solutionsNumber;
+        byte solutionsNumber;
         //readonly List<SudokuTable> solutionTables = new List<SudokuTable>();
         public int TaskId { get { return taskId; } }
-        public int SolutionsNumber { get { return solutionsNumber; } set { solutionsNumber = value; }}
+        public byte SolutionsNumber { get { return solutionsNumber; } set { solutionsNumber = value; }}
         public bool IsMultiSolutions { get { return solutionsNumber >= 2; } }
         public SudokuTable InitialTable { get { return initialTable; } }
